Update convolution.py
Fixing the convLayer. The forwardPass is now able to work properly. Full support to the model has been added, it can be integrated into the model which is provided by the model module. 

Issue to fix is the tradeoff between batch-support and nums of featureMaps. Currently we're implementing batch-support as this way we can integrate it into our model. The best thing i could manage to do in terms of implementing both batch-support and nums of featureMaps is a runtime of O(n^dimOfArray); meaning this is getting progressively worse as we stack more and more convLayers and therefore featureMaps. Once I figure out a viable solution to this problem, I'll be sure to implement it.